updated the feedback footer
added feedback for the package and the documentation!
make important highlight

just switched from manual important to better visibility
added note about frontend proxy

Added a note about using something in front of graylog as recommendation. This way the SSL Setup of Single Host Setups are more easy as this can be done on the frontend like nginx what is realy simple.