clarify on production readiness
removed the part about production readiness in graylog-ctl script description. make no sense at this point as the same text is located in all appliance descriptions too

as i had noticed that a lots of people use the ova in production and did think that it scales automatic just added a few more information to ova in production usage.