added journal messages
After feedback from the community ( https://community.graylog.org/t/graylog-uncommited-messages-deleted-from-journal-utilization-is-too-high/3177/6 ) and some Enterprise Customers I added the error message to the FAQ and put some additional information to it.

This might be good become one chapter "troubleshooting" but currently its better to find that error in the documentation than noting. 

ppl do not read the online description of this error!