add more Info on Upgrade (#395)
* added some more description

I had inclulded some text to have the next steps on update more clear and guide the persons that did not know how to ho handle updates

* removed first sentence

just removed the first sentence that does not make any sense

* added jump marks

just added the jump marks that are needed for the text.

This need to be added to the 2.4 and 3.0 branch as we already have them too (and the versions of that page differs.

* worked on comments

(cherry picked from commit ed280051166b6b1ed0e1bf65919be7a7c4ffcf92)