updates_from_2.4 (#457)
* Updated README  (#452)

* extend information

After I had noticed that the current README was fine for most *dev* people it def. wasn't good for the typical contributor with a not so strong technical backround.

This is the first modification

* rewrote Readme, Updated gitignore

after some testing on windows, rewrote the Readme with the findings and updated the gitignore to reflect the files added on windows.

* reword Readme

Reviewed the Readme

* reviewed changes

* Collector Sidecar: Update table with commands for various init systems (#451)

I think since upstart was only used in Ubuntu which has since switched to systemd since 2015 then it's pretty safe to assume most users aren't using it anymore. Also mostly due to the fact that a .conf file will be required to setup the upstart process so it's not as simple as the other init systems.

Lastly, I know there's a trade off to listing init systems vs distros (before it was distros) but I think this is much better to maintain since some distros recently switched so someone running older version of a distro could be running an init system that isn't specified in the table.


Closes #446

* Fix typo and "the the" (#455)

* Fixing docs (#456)

 - Fixed grammar error in index_model.rst
 - Fixed wrong extractors documentation where dot notation is still being used ( https://community.graylog.org/t/documentation-for-extractors-wrong/5369 )
 - Added CSV generation warning

* Add details about RegExp queries (#447)

* Update queries.rst