Add FAQ section explaining how to enable logging for plugin or specific area of Graylog (#502)
Enabling `debug` logging for the entire Graylog subsystem can generate quite a lot of log messages (making is challenging to view just the desired log messages). Also, this does not affect the logging level for Graylog plugins. 

Add an FAQ section explaining how to enable logging for plugin or specific area of Graylog to address this.

Preview of change:
![screen shot 2018-11-06 at 3 58 42 pm](https://user-images.githubusercontent.com/3423655/48096476-de187680-e1dc-11e8-9673-540cb78df61e.png)