clean up integrations
Included a hidden TOC to have all documents in the TOC and get rid of the warnings. In addition this will display the headlines in the left navigation(!).

Moved the Changelog out of the integrations description page - that can become a really long list that makes it not easy to handle in the future.

created a sub dir for inputs - just to have a nice URL 'integration/inputs/' - what we can use for all other we add in the future. Just to have them seperated and not everything flat.