Adjustments to Integrations documentation (#514)
* moved integrations

* Add Enterprise Integrations installation instructions

* Add explicit download links for packages

* Indicate that all plugins require a license for now

* Adjust wording and fix invalid package names

* Remove unneeded sentence

* fix typo

* More Integrations documentation refinement

- Move to root (no longer just for Enterprise)
- Add Setup page
- Improve overview page

* Minor formatting adjustments

* More formatting adjustments

* clean up integrations

Included a hidden TOC to have all documents in the TOC and get rid of the warnings. In addition this will display the headlines in the left navigation(!).

Moved the Changelog out of the integrations description page - that can become a really long list that makes it not easy to handle in the future.

created a sub dir for inputs - just to have a nice URL 'integration/inputs/' - what we can use for all other we add in the future. Just to have them seperated and not everything flat.

* More revisions to Integrations documentation.

- Change package name to graylog-plugin-enterprise-integrations (open source and no license required, but must use this name since it was built with it).

- Unify Enterprise and open source sections (since we only have open source right now).

- A few other misc changes.

* Revert unintended whitespace changes to conf.py

* Revert unintended whitespace changes to Enterprise Setup

* Another attempt: Revert unintended whitespace changes to Enterprise Setup

* Another attempt: Revert unintended whitespace changes to conf.py

* Fix package name (was missing 'enterprise')

* Remove word "Enterprise" from changelog page for Palo Alto input