added info about python on mac (#746)
After I had to re-initialize my workbench just realized that we do not cover in the readme how you could work with a different python version than the system default one. 

That is the reason I have added a section how to change the python version away from system python. As I only have a Mac to check I firstly added the mac information based on 

https://opensource.com/article/19/5/python-3-default-mac

But that should be easy adoptable for any linux system.