make enterprise more clear (#824)
* added overview of enterprise functions

to make a listing inside the docs what contained in the enterprise plugins. This is created.

* changed function list to hlist

The function hlist makes the complete list of functions a little more compact. 

This will reduce the size of the page.

* Revert "changed function list to hlist"

This reverts commit c7dfab63123462c9529df2c297e88fc83b46c6ae.

* reorder rework text added link

* reworked wording

change order and layout of the listing. Added some more sentences.

* setup page cleanup

cleaned the long list and added the missing download links for the tarball

* reworked after review