created new structure for sending in logs
After I have thought about https://github.com/Graylog2/documentation/issues/147 and how and where to place the information about journald I just realised that we have a really long page. 

Our observation is that shorter pages are more useful for most users, that is why I keep the old page (to not brake navigation) but added sub-pages for the different parts.

Headlines might get some adjustment and all parts can be reworked but that should be the starter.