added overview image & cosmetics
I have added a simple overview image about the message flow and changed the cosmetics a little.

I would love to have all output options at least written for each output method and not only the new available. Because I know that people will not understand that this is inclusive the above list...